ARM: dts: aspeed: Add dts for Supermicro X11
This adds a board definition for the BMC found on a Supermicro X11
series motherboard. The dts was tested on an X11SSH-F board, but is
likely applicable to several boards in this series. The BMC on this
board is an ASPEED AST2400 chip. This dts is not complete (in particular
it is missing all i2c slave devices), but does allow successful boot and
basic function of the BMC on this motherboard.